move around some stuff into the container file
